+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.1 under development
 
-- no changes in this release.
+- Bug #96: Fix unexpected redirects from `Locale` middleware on GET requests (@vjik)
 
 ## 1.0.0 May 22, 2023